// src/app/shared/models/visiting-card.ts
export class VisitingCards {
    id: number;
    roleIds = [];
    template: string;
    dynamicFields: string;
    isActive: boolean;
    isDelete: boolean;
    createdDate: Date;
    modifiedDate: Date;
    createdBy: number;
    modifiedBy: number;
    roleName = [];
    location: string;
    role = []
    constructor() { }
}